components/google_calendar_client_impl/pyproject.toml
[project]
name = "google-calendar-client-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "calendar-client-api",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
]

[tool.uv.sources]
calendar-client-api = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/google_calendar_client_impl"]

[tool.hatch.build.targets.wheel.force-include]
"src/google_calendar_client_impl/py.typed" = "google_calendar_client_impl/py.typed"

[tool.ruff]
line-length = 100
target-version = "py313"
extend = "../../pyproject.toml"

#[tool.ruff.lint]
#ignore = []
